+++ storage-resize-images/functions/src/resize-image.ts
@@ -26,7 +26,25 @@ export function resize(file, size) {
     throw new Error("height and width are not delimited by a ',' or a 'x'");
   }
 
-  return sharp(file, { failOnError: false, animated: config.animated })
+  let sharpOptions = {};
+  try {
+    sharpOptions = JSON.parse(config.sharpOptions);
+  } catch (e) {
+    logs.errorConstuctorOptionsParse(e);
+  }
+
+  /**
+   * Allows customisation of sharp constructor options
+   * Maintains the original config for failOnError
+   * Ensure animated option overrides custom options
+   */
+  const ops = {
+    failOnError: false,
+    ...(sharpOptions || {}),
+    animited: config.animated,
+  };
+
+  return sharp(file, ops)
     .rotate()
     .resize(parseInt(width, 10), parseInt(height, 10), {
       fit: "inside",
@@ -157,10 +175,10 @@ export const modifyImage = async ({
   }
 
   // Path where modified image will be uploaded to in Storage.
-  const modifiedFilePath = path.normalize(
-    config.resizedImagesPath
-      ? path.posix.join(fileDir, config.resizedImagesPath, modifiedFileName)
-      : path.posix.join(fileDir, modifiedFileName)
+  const modifiedFilePath = getModifiedFilePath(
+    fileDir,
+    config.resizedImagesPath,
+    modifiedFileName
   );
   let modifiedFile: string;
 
@@ -169,34 +187,11 @@ export const modifyImage = async ({
 
     // filename\*=utf-8''  selects any string match the filename notation.
     // [^;\s]+ searches any following string until either a space or semi-colon.
-    const contentDisposition =
-      objectMetadata && objectMetadata.contentDisposition
-        ? objectMetadata.contentDisposition.replace(
-            /(filename\*=utf-8''[^;\s]+)/,
-            `filename*=utf-8''${modifiedFileName}`
-          )
-        : "";
-
-    // Cloud Storage files.
-    const metadata: { [key: string]: any } = {
-      contentDisposition,
-      contentEncoding: objectMetadata.contentEncoding,
-      contentLanguage: objectMetadata.contentLanguage,
-      contentType: imageContentType,
-      metadata: objectMetadata.metadata ? { ...objectMetadata.metadata } : {},
-    };
-    metadata.metadata.resizedImage = true;
-    if (config.cacheControlHeader) {
-      metadata.cacheControl = config.cacheControlHeader;
-    } else {
-      metadata.cacheControl = objectMetadata.cacheControl;
-    }
-
-    // If the original image has a download token, add a
-    // new token to the image being resized #323
-    if (metadata.metadata.firebaseStorageDownloadTokens) {
-      metadata.metadata.firebaseStorageDownloadTokens = uuid();
-    }
+    const metadata = constructMetadata(
+      modifiedFileName,
+      imageContentType,
+      objectMetadata
+    );
 
     // Generate a resized image buffer using Sharp.
     logs.imageResizing(modifiedFile, size);
@@ -245,4 +240,60 @@ export const modifyImage = async ({
       logs.errorDeleting(err);
     }
   }
+};
+
+export const constructMetadata = (
+  modifiedFileName: string,
+  imageContentType: string,
+  objectMetadata: ObjectMetadata
+) => {
+  // filename\*=utf-8''  selects any string match the filename notation.
+  // [^;\s]+ searches any following string until either a space or semi-colon.
+  const contentDisposition =
+    objectMetadata && objectMetadata.contentDisposition
+      ? objectMetadata.contentDisposition.replace(
+          /(filename\*=utf-8''[^;\s]+)/,
+          `filename*=utf-8''${modifiedFileName}`
+        )
+      : "";
+
+  // Cloud Storage files.
+  const metadata: { [key: string]: any } = {
+    contentDisposition,
+    contentEncoding: objectMetadata.contentEncoding,
+    contentLanguage: objectMetadata.contentLanguage,
+    contentType: imageContentType,
+    metadata: objectMetadata.metadata ? { ...objectMetadata.metadata } : {},
+  };
+  metadata.metadata.resizedImage = true;
+  if (config.cacheControlHeader) {
+    metadata.cacheControl = config.cacheControlHeader;
+  } else {
+    metadata.cacheControl = objectMetadata.cacheControl;
+  }
+
+  // If the original image has a download token, add a
+  // new token to the image being resized #323
+  if (metadata.metadata.firebaseStorageDownloadTokens) {
+    metadata.metadata.firebaseStorageDownloadTokens = uuid();
+  }
+  return metadata;
+};
+
+const convertToPosixPath = (filePath: string, locale?: "win32" | "posix") => {
+  const sep = locale ? path[locale].sep : path.sep;
+  return filePath.split(sep).join(path.posix.sep);
+};
+export const getModifiedFilePath = (
+  fileDir,
+  resizedImagesPath,
+  modifiedFileName
+) => {
+  return convertToPosixPath(
+    path.posix.normalize(
+      resizedImagesPath
+        ? path.posix.join(fileDir, resizedImagesPath, modifiedFileName)
+        : path.posix.join(fileDir, modifiedFileName)
+    )
+  );
 };